Grafico de cohorte de carreras implementado
