+++ src/services/SubjectDataService.js
@@ -27,7 +27,6 @@ export const getAllSubjectsGuarani = async () => {
 export const getSubjectsByCareer = async (id_carrera) => {
     try {
         const response = await axios.get(`${baseURL}/materias/${id_carrera}`);
-        console.log(`${baseURL}/materias/${id_carrera}`)
         return response;
     } catch (error) {
         return error;
